fix(args): fixed type mismatch in floatVal inside ComputeArgs

+++ silicon-api/src/main/java/org/silicon/computing/ComputeArgs.java
@@ -35,7 +35,7 @@ public class ComputeArgs {
         return this;
     }
 
-    public ComputeArgs floatVal(int value) {
+    public ComputeArgs floatVal(float value) {
         args.add(value);
         return this;
     }
